+++ source/common/http/conn_manager_impl.h
@@ -679,7 +679,7 @@ private:
   const Server::OverloadActionState& overload_disable_keepalive_ref_;
   TimeSource& time_source_;
 
-  const bool validate_header_values_;
+  const bool strict_header_validation_;
 };
 
 } // namespace Http