+++ source/common/http/http1/conn_pool.cc
@@ -343,12 +343,12 @@ void ConnPoolImpl::ActiveClient::onConnectTimeout() {
 
 CodecClientPtr ProdConnPoolImpl::createCodecClient(Upstream::Host::CreateConnectionData& data) {
 
-  const bool validate_header_values =
-      Runtime::runtimeFeatureEnabled("envoy.reloadable_features.validate_header_values");
+  const bool strict_header_validation =
+      Runtime::runtimeFeatureEnabled("envoy.reloadable_features.strict_header_validation");
 
   CodecClientPtr codec{new CodecClientProd(CodecClient::Type::HTTP1, std::move(data.connection_),
                                            data.host_description_, dispatcher_,
-                                           validate_header_values)};
+                                           strict_header_validation)};
   return codec;
 }