+++ source/extensions/filters/network/http_connection_manager/config.cc
@@ -362,18 +362,18 @@ void HttpConnectionManagerConfig::processFilter(
 
 Http::ServerConnectionPtr HttpConnectionManagerConfig::createCodec(
     Network::Connection& connection, const Buffer::Instance& data,
-    Http::ServerConnectionCallbacks& callbacks, const bool validate_header_values) {
+    Http::ServerConnectionCallbacks& callbacks, const bool strict_header_validation) {
   switch (codec_type_) {
   case CodecType::HTTP1:
     return std::make_unique<Http::Http1::ServerConnectionImpl>(
-        connection, callbacks, http1_settings_, maxRequestHeadersKb(), validate_header_values);
+        connection, callbacks, http1_settings_, maxRequestHeadersKb(), strict_header_validation);
   case CodecType::HTTP2:
     return std::make_unique<Http::Http2::ServerConnectionImpl>(
         connection, callbacks, context_.scope(), http2_settings_, maxRequestHeadersKb());
   case CodecType::AUTO:
     return Http::ConnectionManagerUtility::autoCreateCodec(
         connection, data, callbacks, context_.scope(), http1_settings_, http2_settings_,
-        maxRequestHeadersKb(), validate_header_values);
+        maxRequestHeadersKb(), strict_header_validation);
   }
 
   NOT_REACHED_GCOVR_EXCL_LINE;